Moved <BasicCounter> & <Timer> components to TimeComponents.tsx
